feat: refactor button configuration and editor settings in main apps
- Moved the hidden button configuration logic to a new utility file for better organization and reusability.
- Introduced a new function to determine if the macromolecules editor should be disabled, enhancing editor flexibility.
- Updated all main app components (App, ClosableApp, DuoApp, PopupApp) to utilize the new utility functions for managing button visibility and editor settings.
