+++ example/src/App.tsx
@@ -1,27 +1,16 @@
 import { StrictMode, useEffect, useState } from 'react';
-import { ButtonsConfig, Editor, InfoModal } from 'ketcher-react';
+import { Editor, InfoModal } from 'ketcher-react';
 import { Ketcher, StructServiceProvider } from 'ketcher-core';
 
 import 'ketcher-react/dist/index.css';
 
 import { getStructServiceProvider } from './utils';
+import { getHiddenButtonsConfig, isMacromoleculesEditorDisabled } from './utils/editorUrlConfig';
 import { safePostMessage } from './utils/safePostMessage';
-
-const getHiddenButtonsConfig = (): ButtonsConfig => {
-  const searchParams = new URLSearchParams(window.location.search);
-  const hiddenButtons = searchParams.get('hiddenControls');
-
-  if (!hiddenButtons) return {};
-
-  return hiddenButtons.split(',').reduce((acc, button) => {
-    if (button) acc[button] = { hidden: true };
-
-    return acc;
-  }, {} as { [val: string]: { hidden: boolean } });
-};
 
 const App = () => {
   const hiddenButtonsConfig = getHiddenButtonsConfig();
+  const disableMacromoleculesEditor = isMacromoleculesEditorDisabled();
   const [hasError, setHasError] = useState(false);
   const [errorMessage, setErrorMessage] = useState('');
 
@@ -43,6 +32,7 @@ const App = () => {
           setErrorMessage(message.toString());
         }}
         buttons={hiddenButtonsConfig}
+        disableMacromoleculesEditor={disableMacromoleculesEditor}
         staticResourcesUrl={process.env.PUBLIC_URL}
         structServiceProvider={structServiceProvider}
         onInit={(ketcher: Ketcher) => {
